* Add in example merge module
merge-example.cpp: the example, using the future .cpp style of naming
  for windows compatibility.

configure.ac: add in support for enabling building examples.

ExtentField.H: add function to get the name of the field.  Improve
  the error message on setting a non-nullable field.

ExtentSeries.H: [BUGFIX] can not have more records if we don't have an
  extent.

GeneralField.H: [BUGFIX] typo in #define.

Makefile.am: add in the merge example if it is enabled.