[FEATURE] Support string literals and Variable32 comparisons.
The comparison of any pair of string literals or Variable32 fields or
combinations thereof, as well as any field that is convertible to a
string and another argument which is a string literal or Variable32 is
supported with operators ==, !=, >, >=, <, <=.

Add tests for the above.

Stub out scan and parsing support for  =~.  Flames out for now.