* Second checkpoint on the OSR paper; now it's way too long and still      missing sections.
design.tex: Fix up a bunch of the \DataSeries{} bits, add in some extended
  version things.  Properly document a bunch of options we've added lately.

ellard.tex: add label to section so we can refer to it.

programming.tex: Fix up a bunch of the \DataSeries{} bits, add in some extended
  version things.  Fix up the source code, it was slightly old-style.

references.bib: add keys for sorting some of the misc entries.  Add in
  bib entries for the open source location and the tr snapshot.

world-cup-1998.tex: add label to section so we can refer to it.

ExtentType.cpp: Add big comment noting that if you update the source you 
  should update the documentation.