fgrepanalysis.py: 1) Records data points in an SQL database (test) 2) Both prepare and experiment now use no compression, lzf, and gz (so we have a total of five experiments - these three + grep + no-memcpy fgrepanalysis)
experimentation.py: A generic framework for measuring and recording in SQL. This will probably end up somewhere else, but for now it is only used by fgrepanalysis.py.

fgrepanalysis.hg: A simple mercury-plot script that creates a simple, meaningless graph from the points of an experiment. It uses the data from the most recent fgrep experiment.