Added a generic grep module (GrepModule).
Added an analysis for GrepModule (grepanalysis.cpp).

Added an analysis for MemorySortModule and SortModule (sortanalysis.cpp).

Changed the setCurPost/iterator.setPos methods in ExtentSeries so that there it is much faster in non-debug mode.

Added a relocate method to ExtentSeries which sets both the extent and the position at the same time and achieves some performance benefit by doing so.