46/DB - code review of random access to extents
Propose a way to handle nullable fields (& paritally implement)

Propose a way to handle the proliferation of nearly identical code via templated inheritance
(pointer to example code not in checkin)

Commentary on the goal of handling nullable fields with subextentpointer type 1.

Question about unsupportedness of some things in variable32 fields.